Minor Change in README.md
As of now, you can dm Isaiahluke40 IF you have an antivirus that isn't included with the faq. Hugo wont be online for a while but you can still dm him(but he might not respond) so dm me if you want to ask about an antivirus that isn't included.